Se ha escindido la clase class en un fichero independiente. De esta forma que gameoflife sólo contiene la ejecución
